requirements.txt
Flask==3.0.0
Flask_Migrate==4.0.5
flask_sqlalchemy==3.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.22
